+++ default.conf
@@ -1,5 +1,5 @@
 server {
-    listen 80;
+    listen 3000;
     server_name localhost;
 
     root /usr/share/nginx/html;
