+++ src/components/QuizDetails/QuizDetailsView.tsx
