+++ src/components/QuizTitle/QuizTitleView.tsx
