+++ src/components/QuizTry/QuizCollectModal.tsx
