+++ src/components/QuizTry/QuizTryView.tsx
