+++ src/pages/QuizDetails.tsx
@@ -1,210 +1,255 @@
 import React, { useState, useEffect } from 'react';
-import {QuestionDetails} from '../types/QuestionDetails';
 
-import {
-	questionDetailsAdd,
-	questionDetailsDelete,
-	questionDetailsGet, questionDetailsUpdate
-} from '../features/myenglish/MyEnglishAPI';
 
 import { useNavigate, useLocation } from 'react-router-dom';
-import GoToHome from '../components/GoToHome';
-import {QuestionDetailsView} from "../components/QuestionDetailsView";
-import {Modal} from "../components/Modal";
-import _ from "lodash";
+import {UserRepository} from "../repository/UserRepository";
+import {LoginConfirmUserService} from "../application/user/LoginConfirmUserService";
+import {QuizDetailsRepository} from "../repository/QuizDetailsRepository";
+import {UpdateQuizDetailsService} from "../application/quizdetails/UpdateQuizDetailsService";
+import {GetQuizDetailsService} from "../application/quizdetails/GetQuizDetailsService";
+import {RegisterQuizDetailsService} from "../application/quizdetails/RegisterQuizDetailsService";
+import {DeleteQuizDetailsService} from "../application/quizdetails/DeleteQuizDetailsService";
+import {QuizDetailsDTO} from "../dto/QuizDetailsDTO";
+import {QuizDetailsAddModal} from "../components/QuizDetails/QuizDetailsAddModal";
+import {QuizDetailsView} from "../components/QuizDetails/QuizDetailsView";
+import {QuizDetailsEditModal} from "../components/QuizDetails/QuizDetailsEditModal";
 const QuizDetails: React.FC = () =>{
-	// 設問を追加 モーダル表示を制御するtoggle
-	const [isQuestionAddViewModal,setIsQuestionAddViewModal] = useState(false)
-	// 問題を編集 モーダル表示を制御するtoggle
-	const [isQuestionEditViewModal,setIsQuestionEditViewModal] = useState(false)
-	// const [editIndex, setEditIndex] = useState(-1)
 	const navigate = useNavigate();
 	const location = useLocation();
-	const {questionTitle} = location.state || {questionTitle : []};
-	// 問題の設問の配列
-	const [allQuestionDetails,setAllQuestionDetails] = useState<QuestionDetails[]>([]);
-	// 問題の設問と回答オブジェクト
-	const [questionDetails,setQuestionDetails] = useState<QuestionDetails>(
-		{
+	const [isLogin, setIsLogin] = useState(false)
+	const [isShowAddModal, setIsShowAddModal] = useState(false)
+	const [isShowEditModal, setIsShowEditModal] = useState(false)
+	const {quizTitle} = location.state || {quizTitle : []};
+	const [quizDetails, setQuizDetails] = useState<QuizDetailsDTO[]>([])
+	const [editOneQuizDetails, setEditOneQuizDetails] = useState<QuizDetailsDTO>({
+		questionDetailsId : 1,
+		questionTitleId : quizTitle.questionTitleId,
+		questionWord : "",
+		answerCandidateNo1 : "",
+		answerCandidateNo2 : "",
+		answerCandidateNo3 : "",
+		answerCandidateNo4 : "",
+		answerNumber: 1
+	})
+	const [addQuizWord, setAddQuizWord] = useState("")
+	const [addAnswerCandidateNo1, setAddAnswerCandidateNo1] = useState("")
+	const [addAnswerCandidateNo2, setAddAnswerCandidateNo2] = useState("")
+	const [addAnswerCandidateNo3, setAddAnswerCandidateNo3] = useState("")
+	const [addAnswerCandidateNo4, setAddAnswerCandidateNo4] = useState("")
+	const [addAnswerNumber, setAddAnswerNumber] = useState(1)
+	const userRepository =  new UserRepository();
+	const loginConfirmUserService = new LoginConfirmUserService(userRepository);
+	const quizDetailsRepository = new QuizDetailsRepository();
+	const quizDetailsUpdateService = new UpdateQuizDetailsService(quizDetailsRepository);
+	const quizDetailsGetService = new GetQuizDetailsService(quizDetailsRepository);
+	const quizDetailsRegisterService = new RegisterQuizDetailsService(quizDetailsRepository);
+	const quizDetailsDeleteService = new DeleteQuizDetailsService(quizDetailsRepository);
+
+	/**
+	* ログイン確認
+	* 済み(true)	: ログインステータスをログイン済みにする
+	* 未(false)	: トップ画面へリダイレクト
+	* */
+	useEffect(() => {
+		loginConfirmUserService.execute().then(res => {
+			if(!res){
+				redirectQuizTitle();
+				return;
+			}
+			setIsLogin(res);
+		}).catch(() => redirectQuizTitle());
+	}, []);
+
+	/**
+	 * 問題を取得する
+	 * */
+	useEffect(() => {
+		if(isLogin){
+			quizDetailsGetService.execute(quizTitle).then( res => {setQuizDetails(res)});
+		}
+	}, [isLogin]);
+
+	/**
+	 * トップ画面リダイレクト
+	 * */
+	const redirectQuizTitle = () => {
+		navigate("/")
+	}
+
+	/**
+	 * 問題追加モーダルを表示
+	 * */
+	const handleShowAddModal = () => {
+		setIsShowAddModal(true);
+	};
+
+	/**
+	 * 問題追加モーダルを非表示
+	 * */
+	const handleCloseAddModal = () => {
+		setIsShowAddModal(false);
+	};
+
+	/**
+	 * 問題追加入力時の内容を反映する
+	 * 問題追加の初期状態は空
+	 * */
+	const handleChangeAddQuizWord = (event: any) => {
+		setAddQuizWord(event.target.value)
+	}
+	const handleChangeAddAnswerCandidateNo1 = (event: any) => {
+		setAddAnswerCandidateNo1(event.target.value)
+	}
+	const handleChangeAddAnswerCandidateNo2 = (event: any) => {
+		setAddAnswerCandidateNo2(event.target.value)
+	}
+	const handleChangeAddAnswerCandidateNo3 = (event: any) => {
+		setAddAnswerCandidateNo3(event.target.value)
+	}
+	const handleChangeAddAnswerCandidateNo4 = (event: any) => {
+		setAddAnswerCandidateNo4(event.target.value)
+	}
+	const handleChangeAddAnswerNumber = (event: any) => {
+		setAddAnswerNumber(event.target.value)
+	}
+
+	/**
+	 * 問題追加処理
+	 * */
+	const handleAddQuizDetails= () => {
+		const mockQuestionDetailsId : number = 1;
+		quizDetailsRegisterService.execute(
+			mockQuestionDetailsId,
+			quizTitle.questionTitleId,
+			addQuizWord,
+			addAnswerCandidateNo1,
+			addAnswerCandidateNo2,
+			addAnswerCandidateNo3,
+			addAnswerCandidateNo4,
+			addAnswerNumber
+		).catch(() => { console.log("Add Title ERROR") });
+		handleCloseAddModal();
+		window.location.reload(); //サイトを更新して問題情報をサーバと同期する
+	}
+
+	/**
+	 * 問題編集時の内容を反映する
+	 * 問題追加の初期状態は編集対象の問題
+	 * */
+	const handleChangeOneQuizDetails = (event: any) => {
+		const editQuizDetailsElement=event.target.getAttribute("data-key");
+		setEditOneQuizDetails((prevDetail) => ({
+			...prevDetail,
+			[editQuizDetailsElement]:event.target.value
+		}))
+	}
+
+	/**
+	 * 問題を編集するボタンの振る舞い
+	 * @editedOneQuizDetails 編集する問題
+	 * */
+	const handleEditQuizDetails = (editedOneQuizDetails: QuizDetailsDTO) => {
+		setEditOneQuizDetails(editedOneQuizDetails)
+		editShowModal()
+	}
+
+	/**
+	 * 問題を削除するボタンの振る舞い
+	 * */
+	const handleDeleteQuizDetails = (deleteOneQuizDetails: QuizDetailsDTO) => {
+		// TODO: 将来的に画面更新だとレンダリング負荷が高くなりそうなので、表示値変更で値更新としたい
+		quizDetailsDeleteService.execute(deleteOneQuizDetails).catch(() => {console.log("問題の削除でエラーが発生しました")})
+		window.location.reload();
+	}
+
+
+	/**
+	 * 保存ボタンの振る舞い
+	 * 編集された内容をDBに登録する
+	 * */
+	const handleUpdateOneQuizDetails = () => {
+		quizDetailsUpdateService.execute(editOneQuizDetails).catch(() => {console.log("問題の更新でエラーが発生しました")})
+		editCloseModal()
+		// TODO: 将来的に画面更新だとレンダリング負荷が高くなりそうなので、表示値変更で値更新としたい
+		// 問題タイトルと異なる変更方法。モーダルで値を更新する場合は画面更新とする。
+		window.location.reload();
+	}
+
+	/**
+	 * 問題編集モーダルを表示
+	 * */
+	const editShowModal = () => {
+		setIsShowEditModal(true)
+	}
+
+	/**
+	 * 問題編集モーダルを非表示
+	 * */
+	const handleCancelEditModal = () => {
+		const initQuizDetailsDTO: QuizDetailsDTO = {
 			questionDetailsId : 0,
-			questionTitleId : questionTitle.questionTitleId,
+			questionTitleId : quizTitle.questionTitleId,
 			questionWord : "",
 			answerCandidateNo1 : "",
 			answerCandidateNo2 : "",
 			answerCandidateNo3 : "",
 			answerCandidateNo4 : "",
 			answerNumber: 0
-		}
-	);
-	// 問題のタイトルID
-	const questionTitleIdKeyName : string = "questionTitle-"+questionTitle.questionTitleId
-	localStorage.setItem(questionTitleIdKeyName,JSON.stringify(questionTitle));
-
-	// 問題の答え
-	const {questionWord,answerCandidateNo1,answerCandidateNo2,answerCandidateNo3,answerCandidateNo4,answerNumber} = questionDetails;
-	const closeQuestionAddModal = () => {
-		setIsQuestionAddViewModal(false);
+		};
+		setEditOneQuizDetails(initQuizDetailsDTO)
+		editCloseModal()
 	}
 
-	const showQuestionAddModal = () => {
-		setIsQuestionAddViewModal(true);
+	const editCloseModal = () => {
+		setIsShowEditModal(false)
 	}
 
-	// 編集モーダルのcloseがクリックされた時のアクション
-	const closeQuestionEditViewModal = () => {
-		setIsQuestionEditViewModal(false);
+	const handleRedirectQuizTry = () => {
+		navigate('/quiztry',{state : {quizDetails : [...quizDetails]}})
 	}
-
-	// editボタンがクリックされた時のアクション
-	const showQuestionEditViewModal = () => {
-		setIsQuestionEditViewModal(true);
-	}
-
-	/**
- 	* 遷移処理
- 	**/
-	const goToTakeQuiz = () => {
-		console.log(allQuestionDetails)
-		navigate('/takequiz/',{state : {allQuestionDetails : [...allQuestionDetails]}})
-	}
-	const gotoQuizDetailsForm =
-		() => {
-			navigate('/quizdetails/form',{state : {questionTitle : questionTitle}});
-	}
-	const gotoAIQuiz =
-		() => {
-			navigate('/ai',{state : {questionTitle : questionTitle}});
-	}
-
-	/**
-	 * クイズタイトルに紐づくクイズ詳細情報を取得する関数
-	 * @param questionTitle クイズタイトル
-	*/
-	 useEffect(() => {
-		const getAllQuestionDetails = async() => {
-			try{
-				const response = await questionDetailsGet(questionTitle)
-				setAllQuestionDetails(response);
-			}
-			catch(error){
-				alert(error);
-			}
-		}
-		 getAllQuestionDetails();
-	},[])
-
-	const questionDetailsAddHandler = async () => {
-		try{
-			await questionDetailsAdd(questionDetails);
-			closeQuestionAddModal()
-			window.location.reload()
-		}
-		catch(error){
-			alert(error);
-		}
-	};
-
-	/**
-	 * input表示用のonChange関数
-	 * */
-	const detailAddViewForm = (event:any) => {
-		const targetObjectElement=event.target.getAttribute("data-key");
-		setQuestionDetails((prevDetail) => ({
-			...prevDetail,
-			[targetObjectElement]:event.target.value
-		}))
-	}
-
-	// editボタンがクリックされた時の振る舞い
-	const handleEditClick = async (questionDetails:QuestionDetails) =>{
-		try{
-			// update用の問題を登録する
-			const updateQuestionDetail = _.cloneDeep(questionDetails)
-			// 編集用のモーダルを見えるようにする
-			showQuestionEditViewModal()
-			// 編集対象のクイズ詳細を取得するs
-			setQuestionDetails(updateQuestionDetail)
-		}catch(error){
-			alert(error);
-		}
-	}
-
-	const questionDetailsUpdateHandler = async () => {
-		try{
-			await questionDetailsUpdate(questionDetails);
-			closeQuestionAddModal()
-			window.location.reload()
-		}
-		catch(error){
-			alert(error);
-		}
-	};
-
-	const handleDeleteClick = async(details : QuestionDetails) =>{
-		await questionDetailsDelete(details);
-		window.location.reload();
-	}
-
 
 	return (
 		<>
-			<QuestionDetailsView
-				questionDetails={allQuestionDetails}
-				handleEditClick={handleEditClick}
-				handleDeleteClick={handleDeleteClick}
-				handleGotoQuizDetailsForm={gotoQuizDetailsForm}
-				handleGotoAIQuiz={gotoAIQuiz}
-				goToTakeQuiz={goToTakeQuiz}
+			<h1>MyEnglish QuizDetails</h1>
+			{quizDetails.length > 0 ?
+				<>
+					{quizDetails.map((oneQuizDetails: QuizDetailsDTO) => (
+						<QuizDetailsView
+							oneQuizDetails={oneQuizDetails}
+							handleEditQuizDetails={handleEditQuizDetails}
+							handleDeleteQuizDetails={handleDeleteQuizDetails}
+						/>
+					))}
+					<button className="start-button" onClick={handleRedirectQuizTry}>問題に挑戦</button>
+				</>
+				:
+				<p>問題を追加しよう!</p>
+			}
+			<QuizDetailsAddModal
+				isShowModal={isShowAddModal}
+				handleCloseModal={handleCloseAddModal}
+				handleAddQuizDetails={handleAddQuizDetails}
+				addQuizWord={addQuizWord}
+				handleAddQuizWord={handleChangeAddQuizWord}
+				addAnswerCandidateNo1={addAnswerCandidateNo1}
+				handleChangeAddAnswerCandidateNo1={handleChangeAddAnswerCandidateNo1}
+				addAnswerCandidateNo2={addAnswerCandidateNo2}
+				handleChangeAddAnswerCandidateNo2={handleChangeAddAnswerCandidateNo2}
+				addAnswerCandidateNo3={addAnswerCandidateNo3}
+				handleChangeAddAnswerCandidateNo3={handleChangeAddAnswerCandidateNo3}
+				addAnswerCandidateNo4={addAnswerCandidateNo4}
+				handleChangeAddAnswerCandidateNo4={handleChangeAddAnswerCandidateNo4}
+				addAnswerNumber={addAnswerNumber}
+				handleChangeAddAnswerNumber={handleChangeAddAnswerNumber}
 			/>
-			<Modal
-				isViewModal={isQuestionAddViewModal}
-				closeModal={closeQuestionAddModal}
-				viewElements={
-					<>
-						<input placeholder="設問を入力してください" value={questionWord} data-key="questionWord"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補1" value={answerCandidateNo1} data-key="answerCandidateNo1"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補2" value={answerCandidateNo2} data-key="answerCandidateNo2"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補3" value={answerCandidateNo3} data-key="answerCandidateNo3"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補4" value={answerCandidateNo4} data-key="answerCandidateNo4"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="答えの番号を入力" data-key="answerNumber" value={answerNumber}
-							   onChange={detailAddViewForm}></input>
-					</>
-				}
-				requestAPI={questionDetailsAddHandler}
+			<QuizDetailsEditModal
+				isShowModal={isShowEditModal}
+				handleUpdateOneQuizDetails={handleUpdateOneQuizDetails}
+				handleCancelEditModal={handleCancelEditModal}
+				editOneQuizDetails={editOneQuizDetails}
+				handleChangeOneQuizDetails={handleChangeOneQuizDetails}
 			/>
-			<Modal
-				isViewModal={isQuestionEditViewModal}
-				closeModal={closeQuestionEditViewModal}
-				viewElements={
-					//TODO: 問題情報に紐づく回答情報がnullの場合の処理を書く
-					<>
-						<p>編集</p>
-						<input placeholder="設問を入力してください" value={questionWord}
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補1" value={answerCandidateNo1} data-key="answerCandidateNo1"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補2" value={answerCandidateNo2} data-key="answerCandidateNo2"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補3" value={answerCandidateNo3} data-key="answerCandidateNo3"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="回答候補4" value={answerCandidateNo4} data-key="answerCandidateNo4"
-							   onChange={detailAddViewForm}></input>
-						<input placeholder="答えの番号を入力" data-key="answerNumber" value={answerNumber}
-							   onChange={detailAddViewForm}></input>
-					</>
-				}
-				requestAPI={questionDetailsUpdateHandler}
-			/>
-			<div className="under-button-set">
-				<button className="save-button" onClick={showQuestionAddModal}>設問を追加</button>
-				<button className="start-button" onClick={goToTakeQuiz}>Start</button>
-				<GoToHome/>
-			</div>
+			<button className="save-button" onClick={handleShowAddModal}>設問を追加</button>
 		</>
 	)
