--- src/pages/TakeQuiz.tsx
@@ -1,13 +1,0 @@
-import TakeQuizTable from '../components/TakeQuizTable';
-
-const TakeQuiz: React.FC = () =>{
-	/* クイズ出題ページ */
-	return (
-		<div>
-			<h1>Take Question</h1>
-				<TakeQuizTable/>
-		</div>
-	)
-}
-
-export default TakeQuiz;
