--- src/types/QuestionDetails.ts
@@ -1,11 +1,0 @@
-export interface QuestionDetails{
-	/** 詳細なクイズ */
-	questionDetailsId : number;
-	questionTitleId : number;
-	questionWord : string;
-	answerCandidateNo1: string;
-	answerCandidateNo2: string;
-	answerCandidateNo3: string;
-	answerCandidateNo4: string;
-	answerNumber: number;
-}
