--- src/types/QuestionTitle.ts
@@ -1,5 +1,0 @@
-export interface QuestionTitle{
-	questionTitleId : number;
-	ownerUserId :  number;
-	questionTitle : string;
-}
